# Faster Serde Integration (~80% faster)