# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes #.

# Rationale for this change
 
<!--
Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.
-->

Encoding numerics directly in the tape drastically improves the performance of the serde integration.

```
small_i32               time:   [5.3992 µs 5.4006 µs 5.4020 µs]
                        change: [-70.553% -70.532% -70.511%] (p = 0.00 < 0.05)
                        Performance has improved.

large_i32               time:   [5.2606 µs 5.2618 µs 5.2631 µs]
                        change: [-76.768% -76.747% -76.727%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

small_i64               time:   [5.2937 µs 5.2960 µs 5.2986 µs]
                        change: [-73.032% -73.002% -72.974%] (p = 0.00 < 0.05)
                        Performance has improved.

medium_i64              time:   [5.3314 µs 5.3372 µs 5.3417 µs]
                        change: [-77.574% -77.553% -77.533%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) low severe
  1 (1.00%) low mild

large_i64               time:   [5.6473 µs 5.6503 µs 5.6532 µs]
                        change: [-81.152% -81.103% -81.056%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 22 outliers among 100 measurements (22.00%)
  22 (22.00%) low severe

small_f32               time:   [3.6082 µs 3.6101 µs 3.6121 µs]
                        change: [-93.595% -93.591% -93.588%] (p = 0.00 < 0.05)
                        Performance has improved.

large_f32               time:   [3.5233 µs 3.5245 µs 3.5256 µs]
                        change: [-94.058% -94.055% -94.053%] (p = 0.00 < 0.05)
                        Performance has improved.
```

It additionally opens the door to eager parsing in the future, which may yield performance improvements for regular JSON decoding.

I have confirmed this does not regress the performance of the JSON decoder

# What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

# Are there any user-facing changes?


<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!---
If there are any breaking changes to public APIs, please add the `breaking change` label.
-->